+++ app/src/engine/device/LoVoltageInfo.h
@@ -176,52 +176,52 @@ namespace Enercom
         /**
          * Regulator voltage
          */
-        uint32_t regulatorVoltage_{};
+        float regulatorVoltage_{};
 
         /**
          * Minimal regulator voltage
          */
-        uint32_t regulatorVoltageMin_{};
+        float regulatorVoltageMin_{};
 
         /**
          * Maximal regulator voltage
          */
-        uint32_t regulatorVoltageMax_{};
+        float regulatorVoltageMax_{};
 
         /**
          * DAC voltage
          */
-        uint32_t dacVoltage_{};
+        float dacVoltage_{};
 
         /**
          * Minimal DAC voltage
          */
-        uint32_t dacVoltageMin_{};
+        float dacVoltageMin_{};
 
         /**
          * Maximal DAC voltage
          */
-        uint32_t dacVoltageMax_{};
+        float dacVoltageMax_{};
 
         /**
          * Calibration factor
          */
-        uint32_t calibrationFactor_{};
+        float calibrationFactor_{};
 
         /**
          * Calibration offset
          */
-        uint32_t calibrationOffset_{};
+        float calibrationOffset_{};
 
         /**
          * Amplifier ratio
          */
-        uint32_t amplifierRatio_{};
+        float amplifierRatio_{};
 
         /**
          * Regulator proportionality coefficient
          */
-        uint32_t regulatorRatio_{};
+        float regulatorRatio_{};
 
         /**
          * Amplifier attenuator
@@ -231,7 +231,7 @@ namespace Enercom
         /**
          * Required high voltage setting accuracy
          */
-        uint32_t accuracy_{};
+        float accuracy_{};
     };
 
 }
